requirements.txt
scikit-learn
numpy
matplotlib
ipykernel
sentence-transformers
FLAML[automl]
openai
python-dotenv
tiktoken
